в генератор пошли все константы
