--- js/features.js
@@ -1,10 +1,0 @@
-const FEATURES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-];
-
-export {FEATURES};
