--- js/types.js
@@ -1,9 +1,0 @@
-const TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel',
-];
-
-export {TYPES};
